+++ docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Nickel now looks for Cobalt Core in any parent directories relative to its location, before attempting to load from Steam's location.
+* Fixed solo runs removing custom ship starter cards.
 
 ## 1.16.1
 Released 11 May 2025.